requirements.txt
requests==2.32.3
PyYAML==6.0.2
python-dotenv==1.0.1

